requirements.txt
torch
torchvision
transformers
pillow
pandas
bert-score
scikit-learn
tqdm